refactor, fixed security error, layout to mobile
deleted legacy architecture leftovers,
run audit fix 3 times because it generated over 100 issues 1st time,
ended up with just 8 related to actually 2 dep issues.

adjusted layout to mobile

security fix was that I defined a script in public html head,
and the fix is to define it elsewhere and embed it in index.html,
it's suposed to prevent code injection attack.
If there still is a security issue with a dep using string eval()
then I won't be able to fix it. I should have not created the proj with
react create. It seems to be deprecated!